+++ src/classes/Command.js
@@ -22,7 +22,7 @@ module.exports = class Command {
         });
         this.usage = options.optional("usage", null);
         this.disabled = options.optional("disabled", false);
-        this.subcommands = options.optional("subcommands", false);
+        this.subcommands = options.optional("subcommands", null);
         this.cooldown = options.optional("cooldown", 3);
         this.category = options.optional("category", "General");
         if (this.subcommands) {
@@ -50,7 +50,7 @@ module.exports = class Command {
         }
     }
 
-    async prerun(message, t) {
+    async prerun(message, guildDB, t) {
         try {
             await addUser(message.author.id);
         } catch (e) {
@@ -74,19 +74,20 @@ module.exports = class Command {
             const bot_perms = message.guild.me.permissionsIn(message.channel);
 
             if (!bot_perms) {
-                return message.reply(
+                message.reply(
                     `${t("errors:iDontHavePermShort")}\nType \`${
                         guildDB.prefix
                     }help ${
-                        command.name
+                        this.name
                     }\` to get list of permissions required by this command.\nDon't know what you have given already? Type \`${
                         guildDB.prefix
                     }botperms\` in this channel itself.`
                 );
+                return false;
             }
             for (var i = 0; i < basicPerms.length; i++) {
                 if (!bot_perms.has(basicPerms[i])) {
-                    return message.reply(
+                    message.reply(
                         t("errors:iDontHavePermission", {
                             permission: t(
                                 `permissions:${new Permissions(basicPerms[i])
@@ -96,6 +97,7 @@ module.exports = class Command {
                             ),
                         })
                     );
+                    return false;
                 }
             }
         }