+++ dashboard/routes/discord.js
@@ -15,13 +15,13 @@ router.get("/login", (req, res) => {
         res.redirect(
             `https://discord.com/api/oauth2/authorize?client_id=${
                 req.client.user.id
-            }&scope=identify%20guilds&response_type=code&redirect_uri=${
-                `${req.protocol}://${req.get(
-                    "host"
-                )}/discord/callback?redirectUrl=${
-                    encodeURIComponent(req.query.redirectUrl || "/dashboard")
-                }`
-            }`
+            }&scope=identify%20guilds&response_type=code&redirect_uri=${`${
+                req.protocol
+            }://${req.get(
+                "host"
+            )}/discord/callback?redirectUrl=${encodeURIComponent(
+                req.query.redirectUrl || "/dashboard"
+            )}`}`
         );
 });
 //GET /callback
@@ -31,7 +31,12 @@ router.get("/callback", async (req, res) => {
     const params = new URLSearchParams();
     params.set("grant_type", "authorization_code");
     params.set("code", req.query.code);
-    params.set("redirect_uri", `${req.protocol}://${req.get("host")}/discord/callback?redirectUrl=${encodeURIComponent(redirectUrl)}`);
+    params.set(
+        "redirect_uri",
+        `${req.protocol}://${req.get(
+            "host"
+        )}/discord/callback?redirectUrl=${encodeURIComponent(redirectUrl)}`
+    );
     const tokensRes = await fetch("https://discord.com/api/oauth2/token", {
         method: "POST",
         body: params.toString(),
@@ -45,9 +50,7 @@ router.get("/callback", async (req, res) => {
     });
     const tokens = await tokensRes.json();
     if (tokens.error || !tokens.access_token)
-        return res.redirect(
-            `/discord/login?redirectUrl=${redirectUrl}`
-        );
+        return res.redirect(`/discord/login?redirectUrl=${redirectUrl}`);
     const userData = {
         infos: null, //Basic info like user id, tag, username, etc.
         guilds: null,
@@ -56,7 +59,7 @@ router.get("/callback", async (req, res) => {
         if (!userData.infos) {
             response = await fetch("http://discord.com/api/users/@me", {
                 method: "GET",
-                headers: { Authorization: `Bearer ${tokens.access_token}` }
+                headers: { Authorization: `Bearer ${tokens.access_token}` },
             });
             const json = await response.json();
             if (json.retry_after) await req.client.wait(json.retry_after);
@@ -66,7 +69,7 @@ router.get("/callback", async (req, res) => {
         if (!userData.guilds) {
             response = await fetch("http://discord.com/api/users/@me/guilds", {
                 method: "GET",
-                headers: { Authorization: `Bearer ${tokens.access_token}` }
+                headers: { Authorization: `Bearer ${tokens.access_token}` },
             });
             const json = await response.json();
             if (json.retry_after) await req.client.wait(json.retry_after);
@@ -82,15 +85,27 @@ router.get("/callback", async (req, res) => {
     };
     const user = await req.client.users.fetch(req.session.user.id);
     const userDB = await req.client.userDbFuncs.getUser(req.session.user.id);
-    const logsChannel = req.client.channels.cache.get(req.client.config.dashboard.logs);
+    const logsChannel = req.client.channels.cache.get(
+        req.client.config.dashboard.logs
+    );
     if (!userDB.logged && logsChannel && user) {
-        const embed = new Embed({color: "pink"})
-        .setAuthor(user.tag, user.displayAvatarURL())
-        .setDesc(`${user.tag} has logged in to the get dashboard first time ever! :tada:`);
+        const embed = new Embed({ color: "pink" })
+            .setAuthor(user.tag, user.displayAvatarURL())
+            .setDesc(
+                `${user.tag} has logged in to the get dashboard first time ever! :tada:`
+            );
         logsChannel.send({ embeds: [embed] });
-        await req.client.userDbFuncs.updateUser(req.session.user.id, "logged", true);
+        await req.client.userDbFuncs.updateUser(
+            req.session.user.id,
+            "logged",
+            true
+        );
     } else if (!userDB.logged && user) {
-        await req.client.userDbFuncs.updateUser(req.session.user.id, "logged", true);
+        await req.client.userDbFuncs.updateUser(
+            req.session.user.id,
+            "logged",
+            true
+        );
     }
     res.redirect(redirectUrl);
 });