+++ src/config.js
@@ -8,15 +8,21 @@ module.exports = {
     botGuildId: "836854115526770708",
     newsChannelId: "847459283876577360",
     logsChannelId: "855331801635749888",
+    votesChannelId: "867925583777103872",
+    rewardUserOnVote: true,
     suggestionLogsChannelId: "862126837110800414",
     ownerIDs: [
         "815204465937481749" /*PuneetGopinath#0001*/,
         "693754859014324295" /*abhijoshi2k#6842*/,
     ],
+    dbl: {
+        port: 2000,
+        token: process.env.DBL_token,
+    },
     dashboard: {
         port: 8000,
-        secret: process.env.SESS_PASS ?? null,
-        enabled: process.env.SESS_PASS ?? null ? true : false,
-        logs: "855331801635749888"
+        secret: process.env.SESS_SECRET ?? null,
+        enabled: process.env.SESS_SECRET ?? null ? true : false,
+        logs: "855331801635749888",
     },
 };