+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The displayed date is in the format `DD-MM-YYYY`
 ### New features
 
 - refactor(disable cmd): Disable commands using disable cmd now!
+- feat: New cmd `daily` and new category `Economy`
 
 ### Changes