+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The displayed date is in the format `DD-MM-YYYY`
 
 - refactor(disable cmd): Disable commands using disable cmd now!
 - feat: New cmd `daily` and new category `Economy`
+- feat: New cmd `balance`
 
 ### Changes