+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The displayed date is in the format `DD-MM-YYYY`
 
 ### New features
 
+- New cmd `emojify`
 - refactor(disable cmd): Disable commands using disable cmd now!
 - feat: New cmd `daily` and new category `Economy`
 - feat: New cmd `balance`